rfid_monitoring_backend/requirements.txt
Django==4.2.7
pyserial==3.5
